add IOrderDetailService, OrderDetailService

+++ shoappp_backend/shopapp/src/main/java/com/project/shopapp/dtos/OrderDetailDTO.java
@@ -31,7 +31,7 @@ public class OrderDetailDTO {
 
     @JsonProperty("total_money")
     @Min(value = 0, message = "total_money must be >= 0")
-    private Long totalMoney;
+    private Float totalMoney;
 
     private String color;
 }
